LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/at-property-shadow-expected.txt


FAIL @property rules in shadow trees should have no effect assert_equals: expected "calc(1px + 1px)" but got "2px"
